template fixes
